use macro to share common Runtime API impl
