+++ runtime/common/src/lib.rs
@@ -16,4 +16,5 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
+mod apis;
 mod pallets;
